fix(cli): keep the buildx reaper tests off the Windows build
CI cross-compiles this package's test binary for windows/amd64, and
syscall.Kill does not exist there. The POSIX-only tests guarded that with
a runtime.GOOS skip, which never runs because the file fails to compile
first.

Move them to dockerexec_unix_test.go behind //go:build !windows and drop
the now-dead runtime skips. dockerexec_test.go keeps the one test that is
portable. No test coverage changes on Unix.
